Optimizing for mobile.
Instead of running render on mousedown and mousemove it will instead rely
on only the render loop to do that. Doesn't murder performance on mobile
after a few clicks.

Added a stats counter that displays the fps count.

Gotta work on some way to remove blocks on mobile since right click wont
work.